+++ cwt/__init__.py
@@ -16,7 +16,7 @@ from .helpers.hcert import load_pem_hcert_dsc
 from .recipient import Recipient
 from .signer import Signer
 
-__version__ = "1.3.1"
+__version__ = "1.3.2"
 __title__ = "cwt"
 __description__ = "A Python implementation of CWT/COSE"
 __url__ = "https://python-cwt.readthedocs.io"